% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Function to calculate shannon entropy of sequences}
\usage{
calculate_entropy(obj)
}
\arguments{
\item{obj}{named vector obtained from "get_shared_kmers" function}
}
\value{
a list containing a data.frame and a list of plots of shannon values distribution
}
\description{
Function to calculate shannon entropy of sequences
}
\examples{
\dontrun{calculate_entropy(obj)}
}
