% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{count_kmers}
\alias{count_kmers}
\title{Function to extract and count k-mers}
\usage{
count_kmers(obj, kmer_size, threads = 1)
}
\arguments{
\item{obj}{data.frame with a "sample" column with the sample names, and a "merged_path" column, with the full names of the merged files}

\item{kmer_size}{the chosen k-mer size}

\item{threads}{number of threads to use}
}
\value{
a data.frame with a "sample" column, a "count_db" column pointing at .jf files, and a "kmer_txt" column, pointing at the .txt files
}
\description{
Function to extract and count k-mers
}
\examples{
\dontrun{count_kmers(obj, kmer_size = 20, threads = 10)}
}
