% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ed.R
\name{filter_ed}
\alias{filter_ed}
\title{Function to filter based on edit distances}
\usage{
filter_ed(
  x,
  threshold = 3,
  method = "lv",
  chunk_size = 500,
  BPPARAM = BiocParallel::MulticoreParam()
)
}
\arguments{
\item{x}{a list of vectors}

\item{threshold}{minimum number of edits to reach the most similar sequence}
}
\value{
a named list
}
\description{
Function to filter based on edit distances
}
\examples{
\dontrun{(x, threshold = 3, method = "lv", chunk_size  = 500, BPPARAM = MulticoreParam())}
}
