% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_repeated_seqs.R
\name{filter_repeated_seqs}
\alias{filter_repeated_seqs}
\title{Function to filter sequences based on the presence of repeated patterns}
\usage{
filter_repeated_seqs(
  x,
  mode = c("pattern", "any"),
  pattern = "AC",
  min_repeats = 2,
  kmer_length = 3
)
}
\arguments{
\item{x}{a list of vectors}

\item{mode}{used to select if the search should find a pattern (pattern) or any repeated sequence (any)}

\item{pattern}{the pattern to search for if mod = "pattern"}

\item{min_repeats}{the minimum number of repeats to search for}

\item{kmer_length}{used to choose the repeated sequence length if mode = "any"}
}
\description{
Function to filter sequences based on the presence of repeated patterns
}
