% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_shannon_values.R
\name{filter_shannon_values}
\alias{filter_shannon_values}
\title{Function to filter sequences based on their Shannon value}
\usage{
filter_shannon_values(x, threshold)
}
\arguments{
\item{x}{a named vector}
}
\value{
a named vector
}
\description{
Function to filter sequences based on their Shannon value
}
\examples{
\dontrun{filter_shannon_values(x, threshold = 1.8)}
}
