% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared_kmers.R
\name{get_shared_kmers}
\alias{get_shared_kmers}
\title{Function to extract unique k-mers}
\usage{
get_shared_kmers(obj, data, group, min_occurrence)
}
\arguments{
\item{obj}{a data.frame}

\item{data}{a metadata object with a "sample" column}

\item{group}{a column of the metadata to use as grouping factor}

\item{min_occurrence}{minimum number of samples in which k-mers must be found}
}
\value{
a data.frame
}
\description{
Function to extract unique k-mers
}
\examples{
\dontrun{get_shared_kmers(x = obj, data = meta, group= "treatment", min_occurrence = 4)}
}
