% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pairs.R
\name{merge_pairs}
\alias{merge_pairs}
\title{Merge Paired-End FASTQ or FASTA files}
\usage{
merge_pairs(path, pattern_s1, pattern_s2, output_ext)
}
\arguments{
\item{path}{A character string specifying the directory containing the input files. Default is ".".}

\item{pattern_s1}{Pattern to match first-strand (e.g., `*_1.fastq`, `*R1_001.fastq`) files. Default is "*_R1.fastq".}

\item{pattern_s2}{Pattern to match second-strand (e.g., `*_2.fastq`, `*R2_001.fastq`) files. Default is "*_R2.fastq".}

\item{output_ext}{Desired file extension for the output.}
}
\value{
A data.frame with columns `sample` and `merged_path`.
}
\description{
Merges paired-end sequencing files (R1 and R2) in various formats (.fastq, .fq, .fasta, .fa).
}
\examples{
\dontrun{merge_pairs(path = "./path/to/files", pattern_s1 = "_R1.fastq", pattern_s2 = "_R2.fastq", output_ext = "fastq")}
}
