% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shannon.R
\name{calculate_shannon}
\alias{calculate_shannon}
\title{Function to calculate shannon entropy of sequences}
\usage{
calculate_shannon(x)
}
\arguments{
\item{x}{named vector obtained from "get_shared_kmers" function}
}
\value{
a list containing a list of named vectors and a list of plots of shannon values distribution
}
\description{
Function to calculate shannon entropy of sequences
}
\examples{
\dontrun{calculate_entropy(x)}
}
